Extract FunctionListToolCallPromptFormat.
