+++ src/tool/generate-tool-call/index.ts
@@ -2,3 +2,4 @@ export * from "./TextGenerationToolCallModel.js";
 export * from "./ToolCallGenerationEvent.js";
 export * from "./ToolCallGenerationModel.js";
 export * from "./generateToolCall.js";
+export * from "./FunctionListToolCallPromptFormat.js";
